Fix InlineMessage headers/appearance on "no screenshot" dialog
Previously it was in the contentItem because margins were desired, but
that changed in the recent past to make all InlineMessage instances hug
their surrounding items. This one got missed, so the InlineMessage here
was using the header style but floating on the page.


(cherry picked from commit 2c83d9cd7996d12d4281a8d9e8780832d66397bf)

Co-authored-by: Nate Graham <nate@kde.org>