Add checkbox to enable/disable capture delay
This also adds a new captureDelayEnabled setting that defaults to false
and sets the default capture delay to 4 seconds.
I picked 4 because to me 5 feels slightly too long most of the time and 3 feels a bit too short. It's fairly arbitrary though.

We don't want a capture delay by default since that was the previous default.
We do want a default value for capture delay that is greater than zero.
Without that, we'd get a redundant looking set of controls by default
(see below) and it would make the relationship between the controls a bit
less intuitive.

```
[_] Delay: [ No Delay 🡙]
```

Adding the new setting and changing the default setting for the capture delay
also means this should add a migration script. People who used to have capture
delay set to greater than 0 should have capture delay enabled while people who
had it set to 0 should have capture delay disabled and have the capture delay
reset to the default value.