ExportManager: Make Placeholder a bit more efficient
We don't need a dedicated html key since QString::toHtmlEscaped exists and the html key was only used in one place.
We don't need a copy of the key in both the plain text token and the base key member.
We can use a string view instead for the base key member.