ImagePlatformKWin: Use a 60 second timeout for CaptureInteractive
We don't know how long the user will take to pick a window with CaptureInteractive.
A timeout argument of -1 generally gives a 25s timeout.
Let's pick 60s for CaptureInteractive and hope it's enough.
We also need to avoid a potential situation where Spectacle might wait indefinitely for a reply that never comes.