VideoPlatformWayland: Don't use QtConcurrent for creating KPipeWire stream
This was done to hide the latency of querying VAAPI, but means that the
KPipeWire stream lives in the QtConcurrent thread which lacks an event
loop, meaning any event processing the KPipeWire stream does breaks.

Since the KPipeWire stream no longer queries VAAPI on creation, we can
now safely do this on the same thread as the VideoPlatform.