Use declarative registration for most C++ types
This both makes it easier to add C++ types in the future and also makes
the current code a bit clearer by placing the type registration in the
relevant type rather than in the middle of some unrelated C++ file.

Some objects that were registered as singleton instances have been
dropped as they have accessors in other singletons already and are
actually not singletons on the C++ side.