Make it possible to directly instantiate CaptureModeModel
Since we can access the ImagePlatform through SpectacleCore, we can read
the capture mode information from that directly and don't need to inject
it. This in turn makes it possible to create a model instance directly
where needed instead of SpectacleCore needing to create it, reducing the
amount of code and dependencies in SpectacleCore.