Remove unused "show capture instructions" option
Spectacle used to have a giant instructional text box in the Rectangular
Region UI. That was removed in Plasma 6.4, but the UI option to toggle
it remained, and now does nothing.

Rather than applying this UI option to the new very small "Click and
drag to make a section" label, just delete it.

BUG: 507775
FIXED-IN: 6.5.0