Make save and OCR notifications persist in notification history
Apparently the parent of a KNotification must be null for it to remain in history.

I only made these two types remain in history because the notifications about the recording systray icon aren't meant to stay in history.
The copy notifications are technically save notifications, so they are also persistent in history now.